make description of "core.autocrlf" less ambiguous
The description for core.autocrlf refers to reads from / writes to the
"filesystem". While the term is used elsewhere in the config
documentation to refer to the filesystem git is hosted on, it is not
only less clear from context in the case of core.autocrlf, but can also
be plain inaccurate in many cases.

To make more clear the direction of removal / addition of CR when
core.crlf is set, as well as to account for the usage of low-level
commands such as hash-object or cat-file, we change "reading from the
filesystem" to refer instead to "writing to the object database", and
"writing to the filesystem" to "output or writing to the work tree"

Signed-off-by: Will Palmer <wmpalmer@gmail.com>