refactor pretty.c into "parse" and "format" steps
Previously we parsed-out the user-defined "pretty" formats and outputted
the result in one pass, once for every commit. This was perfectly
reasonable, as the performance penalties for doing so are almost
nonexistant. However, it did make it cumbersome to define more-complex
placeholders.

Here we split the bulk of pretty.c into separate "parse" and "format"
steps, adding the bulk of the back-end for more-complicated formatting
rules in the process.

This also adds test-pretty, for debugging the parser. It prints out a
human-readable dump of the parsed format.

Signed-off-by: Will Palmer <wmpalmer@gmail.com>