refactor format_person_part to parse_commit_person
format_person_part was closely tied to the expansion of a user format's
placeholders. This is no good if we want to split the user format
processing into separate "parsing" and "processing" stages.

parse_commit_person is introduced as the "processing" part of
format_person_part, concerning itself only with the format of the commit
author or committer, without worrying about which part might be desired.

This does mean we lose the small benefit of being able to skip the
finding of the date-part if we want something which comes before it, but
in exchange for doing so, we get a structure which can potentially be
re-used for multiple placeholders (eg, formats with both %an and %ad).

There is a magic number in this commit, providing a sanity-check on the
length of a name or mail part. The number is taken from fmt_ident in
ident.c, and while it is technically smaller than what can be considered
okay in this context (ie, the buffer we're working with will be of size
1024, not 1000), it makes sense to be consistent about what we consider
to be valid.

Signed-off-by: Will Palmer <wmpalmer@gmail.com>