(WIP) generic signing-statements for commits
--signoff doesn't make much sense on the face of it: It's a shortcut
for adding a message which was, at the time, specific to a single
project. Other projects have sense adopted the practice as well, but
often (if questions in #git are any indication) due to
misunderstandings as to what it really means (or, in fact, the
mistaken belief that it has any meaning at all)

The idea of --signoff is useful, but --signoff itself is not. What
is required instead is a generic method for adding "signing statements"
to commits. This becomes even more obvious when you note the siblings
of the Signed-off-by: line which have sprung up. Notably:
	CC:
	Acked-by:
	Reviewed-by:
	Tested-by:

Here we attempt to provide a generic interface for defining signing
statements, including the addition of commit.signoff.<statement> for
user-defined statements.