fix potential overflow in mailmapped formats
this corrects a potential overflow when setting the "end" of a
mailmapped name/mail. If *mail_start points to a high-enough address,
and the mailmapped person_mail is long enough, the calculation:
    name_end = name_start + strlen(person_name);
could theoretically wrap.

Though, honestly, the real reason is because it looked weird to me and
was hard to read when I was refactoring it. The "end" was generated by
adding the strlen to the "start", but its only purpose was to calculate
the length of the name/email anyway.

Here we attempt to restore sanity by adding {name,email}_len, and using
that instead.

Signed-off-by: Will Palmer <wmpalmer@gmail.com>