interpret %C(invalid) as we would %%C(invalid)
%C(...) had the distinction of being the only format placeholder which
could trigger a die(), a side-effect of its ancestry in calling the
existing color_parse_mem(...). This was good, because it gave users an
explanation of what went wrong. It was also inconsistent, since every
other "unknown placeholder" was interpreted as a literal.

This removes the inconsistency by interpreting %C(invalid) as a literal.
Perhaps this is a step in the wrong direction.

Signed-off-by: Will Palmer <wmpalmer@gmail.com>