add long forms for author and committer identity
this adds the long forms %(authorname), %(committername),
%(authoremail), and %(committeremail) to complement the existing
placeholders %an, %ae, %cn, and %ce; and, :mailmap forms of each, eg:
    %(authoremail:mailmap)

The main purpose of this change is to match the format of the
placeholders supported in git for-each-ref, though the optional :mailmap
parameter seemed like a sensible extension.

At this point we have enough "long form" placeholders to justify giving
them their own section in the documentation, so we do.

Signed-off-by: Will Palmer <wmpalmer@gmail.com>