teach command-listing components about builtins
'git help -a' and the "unknown command lookup" both searched for
potential matches via the $PATH and GIT_EXEC_PATH. While this behavior
produced the expected results most of the time, it was not technically
correct, as the majority of the "dashed" commands are mere wrappers for
built-in commands at this point. Setting the GIT_EXEC_PATH to something
other than the "true" path could result in git reporting that there were
no commands available, and using the "help" command to do so. Teaching
these paths about builtins would also serve the general trend of moving
away from the "dashed" forms of commands.

To correct this, we make the "builtin command list" used by git.c
available to other builtins, then use that to build a complete list of
available commands.

as it should no longer be possible for git to have no commands available,
thei "Uh oh." warning during unknown-command searching is removed.