core.extrapath for alternative sources of commands
prior to this change, there were really only two options for adding
additional git-* commands: adding the scripts to git's libexec dir, or
adding additional directories to your regular PATH.

Both of these options are workable, but sub-optimal. libexec is often
in a shared location, only writeable by root, while cluttering PATH with
git-specific commands subverts the purpose of having "libexec" in the
first place. Really, this is a configuration issue, so it should be
handled via the git config.

We add core.extrapath, which can be specified multiple times, each time
telling git about another place in which it can look for git-*
executables. The order of preference is now:
	1) builtin commands
	2) dashed commands in libexec
	3) core.extrapath
	4) aliases

As loading the config file requires setup..() having already occured, we
move its invocation to outside of handle_alias(). This further requires
some dancing around with chdir() due to some conflicting assumptions
made regarding "where we are when reading the config" and "where we are
when exec()ing to an external program"